Moved core.next.storagemanager.Dataset to core.next.query
