+++ src/main/java/fr/inria/corese/core/next/query/api/Operation.java
@@ -1,8 +1,8 @@
 package fr.inria.corese.core.next.query.api;
 
 import fr.inria.corese.core.next.data.api.Value;
-import fr.inria.corese.core.next.storagemanager.api.dataset.Dataset;
 import fr.inria.corese.core.next.query.api.result.BindingSet;
+import fr.inria.corese.core.next.query.dataset.Dataset;
 
 /**
  * A generic SPARQL operation (query or update) on a Corese repository.
