+++ src/main/java/fr/inria/corese/core/next/query/api/repository/RepositoryConnection.java
@@ -1,9 +1,9 @@
 package fr.inria.corese.core.next.query.api.repository;
 
 import fr.inria.corese.core.next.data.api.ValueFactory;
-import fr.inria.corese.core.next.storagemanager.api.dataset.Dataset;
 import fr.inria.corese.core.next.query.api.exception.QuerySyntaxException;
 import fr.inria.corese.core.next.query.api.exception.RepositoryException;
+import fr.inria.corese.core.next.query.dataset.Dataset;
 import fr.inria.corese.core.next.query.api.*;
 
 /**
